Split bleDatalogger.ts into per-protocol modules under src/lib/ble/
Behavior-preserving refactor. The 1030-line monolith is gone — each of
the four BLE protocols now lives in its own file, and `bleDatalogger.ts`
becomes a one-line backward-compat barrel.

New structure (src/lib/ble/):
  index.ts          (33)   Public API barrel
  types.ts          (36)   BleConnection, FileInfo, DownloadProgress, BatteryInfo
  internal.ts       (17)   UUIDs + debug logger (NOT re-exported)
  format.ts         (32)   formatBytes / formatSpeed / formatTime
  connection.ts     (58)   isBleSupported, connectToDevice, disconnect
  fileTransfer.ts   (276)  LIST + GET protocol (data-log download) + parseFileList
  battery.ts        (66)   BATT protocol
  settings.ts       (262)  SLIST/SGET/SSET/SRESET protocol
  trackSync.ts      (380)  TLIST/TGET/TPUT/TDEL track-file protocol

  src/lib/bleDatalogger.ts (11)  thin `export * from "./ble";` shim

No file is over 380 lines now (was 1030). Adding a new BLE command no
longer means appending to a 1000-line file with one shared notification
handler — it lives next to its protocol siblings.

What did NOT change:
  - All public function signatures, types, and observable behavior.
    Each function moved verbatim, just into a new file.
  - The 11 `new Promise(async ...)` executors. Each still has the
    accompanying eslint-disable comment explaining why the antipattern
    is contained (inner try/catch handles rejection). The cleaner IIFE
    rewrite was deliberately not bundled with this PR — keeping the
    change a pure structural move makes review and revert trivial.
  - All 6 import sites (DeviceContext, DataloggerDownload,
    FileManagerDrawer, DeviceTracksTab, DeviceSettingsTab) still
    import from `@/lib/bleDatalogger`. The legacy barrel re-exports
    everything so they keep working unchanged.

CLAUDE.md architecture map updated to reflect the new directory layout.

What this unlocks:
  Now that each protocol is isolated, the next PR can add Vitest tests
  per protocol with a small mock of `BluetoothRemoteGATTCharacteristic`
  (just `startNotifications`, `addEventListener`, `writeValue`,
  `removeEventListener`). The previous monolith was hard to test
  because mocking one shared notification handler meant simulating
  every protocol's wire format simultaneously.

Verified:
  - npm run lint clean
  - npm run typecheck (tsc -b) clean
  - npm run test:run (203 passed)
  - npm run build clean

Next PR: BLE protocol tests built on top of this structure.
